+++ backend/src/main/java/com/ctbcbank/lab/backend/model/response/CustomerAccountResponse.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -13,5 +15,5 @@ public class CustomerAccountResponse {
 
     private Integer id;
     private String name;
-    private Double valueAccount;
+    private BigDecimal valueAccount;
 }
